# shared/api/openapi.json
{"openapi":"3.1.0","info":{"title":"api.learn.framer.university","description":"API documentation for [learn.framer.university](https://learn.framer.university)","contact":{"name":"Sakibul Islam","email":"sakibulislam25800@gmail.com"},"license":{"name":""},"version":"0.0.0"},"servers":[{"url":"https://learn.framer.university"},{"url":"https://staging.learn.framer.university"},{"url":"http://localhost:8080"}],"paths":{"/v1":{"get":{"summary":"Health check","operationId":"health_check","responses":{"200":{"description":"successful operation","content":{"application/json":{"schema":{"$ref":"#/components/schemas/MessageResponse"}}}}}}},"/v1/auth/continue/{token}":{"get":{"tags":["auth"],"summary":"Verify the user's email","operationId":"continue_signin","parameters":[{"name":"email_token","in":"path","description":"Token used to verify email","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"successful operation","content":{"application/json":{"schema":{"$ref":"#/components/schemas/VerifiedEmailResponse"}}}},"400":{"description":"Verification token has expired","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}}}}},"/v1/auth/signin":{"post":{"tags":["auth"],"summary":"Sign in the user and send a sign-in email","operationId":"signin","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AuthSignInBody"}}},"required":true},"responses":{"200":{"description":"successful operation","content":{"application/json":{"schema":{"$ref":"#/components/schemas/MessageResponse"}}}}}}}},"components":{"schemas":{"AuthSignInBody":{"type":"object","required":["email"],"properties":{"email":{"type":"string"}}},"ErrorResponse":{"type":"object","required":["title","status","detail"],"properties":{"detail":{"type":"string"},"status":{"type":"integer","format":"int32","minimum":0},"title":{"type":"string"}}},"MessageResponse":{"type":"object","required":["message"],"properties":{"message":{"type":"string"}}},"VerifiedEmailResponse":{"type":"object","required":["access_token","refresh_token"],"properties":{"access_token":{"type":"string"},"refresh_token":{"type":"string"}}}},"securitySchemes":{"bearer":{"type":"http","scheme":"bearer","bearerFormat":"JWT"}}}}